refactor(MailboxManager): Split into category dedicated files
