+++ MailCore/API/MailApiFetcher/MailApiFetcher+Common.swift
