+++ MailCore/API/MailApiFetcher/MailApiFetcher+Extended.swift
