+++ MailCore/API/NetworkRequestRetrier.swift
@@ -1,0 +1,74 @@
+/*
+ Infomaniak Mail - iOS App
+ Copyright (C) 2022 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Foundation
+import Alamofire
+
+final class NetworkRequestRetrier: RequestInterceptor {
+    let maxRetry: Int
+    private var retriedRequests: [String: Int] = [:]
+    let timeout = -1001
+    let connectionLost = -1005
+
+    init(maxRetry: Int = 3) {
+        self.maxRetry = maxRetry
+    }
+
+    func retry(
+        _ request: Alamofire.Request,
+        for session: Session,
+        dueTo error: Error,
+        completion: @escaping (RetryResult) -> Void
+    ) {
+        guard request.task?.response == nil,
+              let url = request.request?.url?.absoluteString else {
+            removeCachedUrlRequest(url: request.request?.url?.absoluteString)
+            completion(.doNotRetry)
+            return
+        }
+
+        let errorGenerated = error as NSError
+        switch errorGenerated.code {
+        case timeout, connectionLost:
+            guard let retryCount = retriedRequests[url] else {
+                retriedRequests[url] = 1
+                completion(.retryWithDelay(0.5))
+                return
+            }
+
+            if retryCount < maxRetry {
+                retriedRequests[url] = retryCount + 1
+                completion(.retryWithDelay(0.5))
+            } else {
+                removeCachedUrlRequest(url: url)
+                completion(.doNotRetry)
+            }
+
+        default:
+            removeCachedUrlRequest(url: url)
+            completion(.doNotRetry)
+        }
+    }
+
+    private func removeCachedUrlRequest(url: String?) {
+        guard let url else {
+            return
+        }
+        retriedRequests.removeValue(forKey: url)
+    }
+}
